AI_Accelerated_Image_Recognition_Project/requirements.txt
# Core
flask==3.0.3
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
numpy==1.26.4
gunicorn==21.2.0

# Optional (if you add CORS or logging middlewares later)
# flask-cors
